requirements.txt
langchain
python-dotenv
streamlit
streamlit_extras
openai
qdrant-client
tiktoken
ebooklib
google-api-python-client
wikipedia
pytesseract
Pillow


# uncomment to use huggingface llms
# huggingface-hub==0.14.1

# uncomment to use instructor embeddings
# InstructorEmbedding==1.0.1
# sentence-transformers==2.2.2